requirements.txt
streamlit
requests
pefile
pyelftools
